tools/requirements.txt
playwright==1.44.0
supabase==2.4.3
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
